requirements.txt
asgiref==3.8.1
Django==5.0.4
django-model-utils==4.5.0
django-notifications-hq==1.8.3
django-push-notifications==3.0.2
jsonfield==3.1.0
lxml==5.2.1
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
signals==0.0.2
sqlparse==0.4.4
swapper==1.3.0
